+++ docs/changelog.rst
@@ -4,6 +4,13 @@ Changelog
 Below you'll find all the changes that have been made to the code with
 newest changes first.
 
+0.16.x
+------
+
+* v0.16.0
+    * Added ability to detect if sample program is an image and return image type
+    * Added ability to get path to sample program
+
 0.15.x
 ------